+++ src/app/admin/beneficiaries/beneficiary-edit/beneficiary-edit.component.ts
@@ -7,7 +7,7 @@ import { MatSnackBar } from '@angular/material/snack-bar';
 
 import { BeneficiariesFacade } from '../beneficiaries.facade';
 import { Beneficiary } from '@shared/models';
-import { KIV_ZONES, SPECIAL_CONDITIONS } from '../../../constants';
+import { KIV_ZONES, SPECIAL_CONDITIONS } from '@shared/constants';
 import { COMMON_FIELDS } from '../beneficiary-new/beneficiary-new.component';
 
 @Component({
