+++ src/app/admin/volunteers/volunteers-details/volunteers-details.component.ts
@@ -21,7 +21,7 @@ import { MatAutocompleteSelectedEvent } from '@angular/material/autocomplete';
 import { MatDialog } from '@angular/material/dialog';
 import { EsriMapComponent } from '@shared/esri-map/esri-map.component';
 import { MatSnackBar } from '@angular/material/snack-bar';
-import { KIV_ZONES } from '../../../constants';
+import { KIV_ZONES } from '@shared/constants';
 
 const minTemp = 36;
 const maxTemp = 41;
